pyproject.toml
[tool.poetry]
name = "flask-rag-service"
version = "0.1.0"
description = "Web application and REST API Retrieval Augmented Generation microservice with distributed task preprocessing and vector database"
authors = ["Michael Conan <michaelbconan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
celery = {extras = ["redis"], version = "^5.4.0"}
Flask-SQLAlchemy = "^3.1.1"
flasgger = "^0.9.7.1"
psycopg2-binary = "^2.9.9"
flask-jwt-extended = "^4.6.0"
flask-bcrypt = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
black = "^24.4.2"
bandit = "^1.7.9"
flake8 = "^7.1.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
